внес правки 1.0

+++ src/main/java/org/practice/bank/dto/transaction/TransactionDto.java
@@ -15,5 +15,5 @@ public record TransactionDto (
         Account accountTo,
         LocalDate date,
         Time time,
-        Long sum
+        Double sum
 ) {}
